requirements.txt
openai>=1.0.0
openai-whisper
faster-whisper
ffmpeg-python
pydub
python-dotenv
numpy<2
